+++ webots_ros2_core/CHANGELOG.rst
@@ -5,7 +5,7 @@ Changelog for package webots_ros2_core
 1.0.0 (2019-XX-YY)
 ------------------
 * Added support for multi robots.
-* Added a new tf_publisher node to publish tf of desired Webots nodes.
+* Added a new TfPublisher class to publish transforms of all the Solid nodes of the robot (if the robot `supervisor` field is true).
 * Added the possibility to run nodes in synchronized mode (using the 'synchronization' parameter).
 * Added a laser_publisher module.
 * Added better support for differential drive robots ('WebotsDifferentialDriveNode' class).